pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-presentation-agent"
version = "0.1.0"
description = "Automated slide generation agent system scaffold."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Auto Presentation Agent Authors" }]
dependencies = [
    "pandas>=2.2",
    "python-pptx>=0.6.23",
]

[project.scripts]
auto-presentation-agent = "auto_presentation_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]
